remove optional classname from jsonschemafield and remove old date and location jsx
