+++ src/scenes/Moves/Ppm/DateAndLocation.jsx
@@ -1,154 +1,113 @@
-import { pick, get } from 'lodash';
+import { get, pick } from 'lodash';
 import PropTypes from 'prop-types';
-import React, { Fragment } from 'react';
+import React, { Component, Fragment } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { reduxForm } from 'redux-form';
-import {
-  renderField,
-  recursivelyAnnotateRequiredFields,
-  validateRequiredFields,
-  addUiSchemaRequiredFields,
-} from 'shared/JsonSchemaForm';
-import WizardPage from 'shared/WizardPage';
+import { getFormValues } from 'redux-form';
 import YesNoBoolean from 'shared/Inputs/YesNoBoolean';
 import { loadPpm, createOrUpdatePpm } from './ducks';
+import { reduxifyWizardForm } from 'shared/WizardPage/Form';
+import { SwaggerField } from 'shared/JsonSchemaForm/JsonSchemaField';
+
 import './DateAndLocation.css';
 
-const NULL_ZIP = ''; //HACK: until we can figure out how to unset zip
 const formName = 'ppp_date_and_location';
-const subsetOfFields = [
-  'planned_move_date',
-  'pickup_postal_code',
-  'destination_postal_code',
-  'additional_pickup_postal_code',
-  'days_in_storage',
-];
+const DateAndLocationWizardForm = reduxifyWizardForm(formName);
 
-export class DateAndLocation extends React.Component {
-  state = { showAdditionalPickup: false, showTempStorage: false };
-  static getDerivedStateFromProps(nextProps, prevState) {
-    const result = {};
-    if (
-      get(
-        nextProps,
-        'formData.values.additional_pickup_postal_code',
-        NULL_ZIP,
-      ) !== NULL_ZIP
-    )
-      result.showAdditionalPickup = true;
-    if (get(nextProps, 'formData.values.days_in_storage', 0) > 0)
-      result.showTempStorage = true;
-    return result;
-  }
-  setShowAdditionalPickup = show => {
-    this.setState({ showAdditionalPickup: show }, () => {
-      if (!show) this.props.change('additional_pickup_postal_code', NULL_ZIP);
-    });
-  };
-  setShowTempStorage = show => {
-    this.setState({ showTempStorage: show }, () => {
-      if (!show) this.props.change('days_in_storage', '0');
-    });
-  };
+export class DateAndLocation extends Component {
   componentDidMount() {
     document.title = 'Transcom PPP: Date & Locations';
     const moveId = this.props.match.params.moveId;
     this.props.loadPpm(moveId);
   }
   handleSubmit = () => {
-    const { createOrUpdatePpm, dirty } = this.props;
-    if (dirty) {
+    const pendingValues = Object.assign({}, this.props.formValues);
+    if (pendingValues) {
+      pendingValues['has_additional_postal_code'] =
+        pendingValues.has_additional_postal_code || false;
+      pendingValues['has_sit'] = pendingValues.has_sit || false;
       const moveId = this.props.match.params.moveId;
-      const pendingValues = Object.assign({}, this.props.formData.values);
-      //HACK: temp work around until we figure out how to unset additional_pickup_postal_code
-      if (pendingValues.additional_pickup_postal_code === NULL_ZIP)
-        delete pendingValues.additional_pickup_postal_code;
-      createOrUpdatePpm(moveId, pendingValues);
+      this.props.createOrUpdatePpm(moveId, pendingValues);
     }
   };
   render() {
     const {
-      schema,
       pages,
       pageKey,
-      valid,
-      dirty,
       hasSubmitSuccess,
       error,
+      currentOrders,
+      currentPpm,
+      initialValues,
     } = this.props;
-    const { showAdditionalPickup, showTempStorage } = this.state;
-    const uiSchema = {
-      requiredFields: [
-        'planned_move_date',
-        'pickup_postal_code',
-        'destination_postal_code',
-      ],
-    };
-    if (showAdditionalPickup)
-      uiSchema.requiredFields.push('additional_pickup_postal_code');
-    if (showTempStorage) uiSchema.requiredFields.push('days_in_storage');
-    addUiSchemaRequiredFields(schema, uiSchema);
-    recursivelyAnnotateRequiredFields(schema);
-    const fields = schema.properties || {};
     return (
-      <WizardPage
+      <DateAndLocationWizardForm
         handleSubmit={this.handleSubmit}
-        isAsync={true}
         pageList={pages}
         pageKey={pageKey}
-        pageIsValid={valid}
-        pageIsDirty={dirty}
         hasSucceeded={hasSubmitSuccess}
-        error={error}
+        serverError={error}
+        initialValues={initialValues}
+        enableReinitialize={true} //this is needed as the pickup_postal_code value needs to be initialized to the users residential address
       >
-        <form>
-          <h1 className="sm-heading">PPM Dates & Locations</h1>
-          <h3> Move Date </h3>
-          {renderField('planned_move_date', fields, '')}
-          <h3>Pickup Location</h3>
-          {renderField('pickup_postal_code', fields, '')}
-          <p>Do you have stuff at another pickup location?</p>
-          <YesNoBoolean
-            value={showAdditionalPickup}
-            onChange={this.setShowAdditionalPickup}
-          />
-          {this.state.showAdditionalPickup && (
-            <Fragment>
-              {renderField('additional_pickup_postal_code', fields, '')}
-              <p>Making additional stops may decrease your PPM incentive.</p>
-            </Fragment>
-          )}
-          <h3>Destination Location</h3>
-          <p>
-            Enter the ZIP for your new home if you know it, or for{' '}
-            {this.props.currentOrders &&
-              this.props.currentOrders.new_duty_station.name}{' '}
-            if you don't.
-          </p>
-          {renderField('destination_postal_code', fields, '')}
-          The ZIP code for{' '}
+        <h1 className="sm-heading">PPM Dates & Locations</h1>
+        <h3> Move Date </h3>
+        <SwaggerField
+          fieldName="planned_move_date"
+          swagger={this.props.schema}
+          required
+        />
+        <h3>Pickup Location</h3>
+        <SwaggerField
+          fieldName="pickup_postal_code"
+          swagger={this.props.schema}
+          required
+        />
+        <SwaggerField
+          fieldName="has_additional_postal_code"
+          swagger={this.props.schema}
+          component={YesNoBoolean}
+        />
+        {get(this.props, 'formValues.has_additional_postal_code', false) && (
+          <Fragment>
+            <SwaggerField
+              fieldName="additional_pickup_postal_code"
+              swagger={this.props.schema}
+              required
+            />
+            Making additional stops may decrease your PPM incentive.
+          </Fragment>
+        )}
+        <h3>Destination Location</h3>
+        <p>
+          Enter the ZIP for your new home if you know it, or for{' '}
           {this.props.currentOrders &&
             this.props.currentOrders.new_duty_station.name}{' '}
-          is{' '}
-          {this.props.currentOrders &&
-            this.props.currentOrders.new_duty_station.address.postal_code}{' '}
-          <p>
-            Are you going to put your stuff in temporary storage before moving
-            into your new home?
-          </p>
-          <YesNoBoolean
-            value={showTempStorage}
-            onChange={this.setShowTempStorage}
-          />
-          {this.state.showTempStorage && (
-            <Fragment>
-              {renderField('days_in_storage', fields, '')}
-              <p>You can choose up to 90 days.</p>
-            </Fragment>
-          )}
-        </form>
-      </WizardPage>
+          if you don't.
+        </p>
+        <SwaggerField
+          fieldName="destination_postal_code"
+          swagger={this.props.schema}
+          required
+        />
+        The ZIP code for {currentOrders && currentOrders.new_duty_station.name}{' '}
+        is {currentOrders && currentOrders.new_duty_station.address.postal_code}{' '}
+        <SwaggerField
+          fieldName="has_sit"
+          swagger={this.props.schema}
+          component={YesNoBoolean}
+        />
+        {get(this.props, 'formValues.has_sit', false) && (
+          <Fragment>
+            <SwaggerField
+              fieldName="days_in_storage"
+              swagger={this.props.schema}
+              required
+            />{' '}
+            You can choose up to 90 days.
+          </Fragment>
+        )}
+      </DateAndLocationWizardForm>
     );
   }
 }
@@ -157,7 +116,6 @@ DateAndLocation.propTypes = {
   schema: PropTypes.object.isRequired,
   loadPpm: PropTypes.func.isRequired,
   createOrUpdatePpm: PropTypes.func.isRequired,
-  currentServiceMember: PropTypes.object,
   error: PropTypes.object,
   hasSubmitSuccess: PropTypes.bool.isRequired,
 };
@@ -174,15 +132,15 @@ function mapStateToProps(state) {
     currentOrders:
       get(state.loggedInUser, 'loggedInUser.service_member.orders[0]') ||
       get(state.orders, 'currentOrders'),
-    formData: state.form[formName],
-    enableReinitialize: true,
+    currentPpm: get(state.ppm, 'currentPpm'),
+    formValues: getFormValues(formName)(state),
   };
   const defaultPickupZip = get(
     state.loggedInUser,
     'loggedInUser.service_member.residential_address.postal_code',
   );
   props.initialValues = props.currentPpm
-    ? pick(props.currentPpm, subsetOfFields)
+    ? props.currentPpm
     : defaultPickupZip
       ? {
           pickup_postal_code: defaultPickupZip,
@@ -194,10 +152,4 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators({ loadPpm, createOrUpdatePpm }, dispatch);
 }
 
-const DateAndLocationForm = reduxForm({
-  form: formName,
-  validate: validateRequiredFields,
-})(DateAndLocation);
-export default connect(mapStateToProps, mapDispatchToProps)(
-  DateAndLocationForm,
-);
+export default connect(mapStateToProps, mapDispatchToProps)(DateAndLocation);
